+++ backend/database/find.js
@@ -10,3 +10,4 @@ const find = async () => {
 }
 
 module.exports = find;
+
